requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-pinecone>=0.1.0
openai>=1.12.0
pinecone-client>=3.0.0
streamlit>=1.32.0
pandas>=2.2.0
matplotlib>=3.8.0
python-dotenv>=1.0.0 